+++ changelog.md
@@ -2,6 +2,7 @@ v1.1.0:
  - added forcable x86 and arm, no matter the platform (so arm can emulate an x86 environ and vice versa)
  - added builds for cross platform and cross architecture
  - added notarization for macos builds 
+ - fixed upgrade command not properly checking docker hub
 
 
  notes on v1.1.0: